Cast all register offsets to their 8-bit limit
The argument to reg_8, reg_16, and reg_32 is an
8-bit register offset.

The only variables passed to these exclusively hold
8-bit values, too:
- inst.rm_index,
- inst.rm_eai, and
- inst.code.extra

So it would make sense to switch them to uint8_t, however
they are unfortunately Bitu's with non-trivial unions,
so for now we'll simply cast them when passed
to the register offsets.

+++ src/cpu/core_full/op.h
@@ -615,7 +615,7 @@ switch (inst.code.op) {
 		if (CPU_ArchitectureType<ArchitectureType::Intel486NewSlow) goto illegalopcode;
 		FillFlags();
 		if (inst_op1_d==reg_eax) {
-			inst_op1_d=reg_32(inst.rm_index);
+			inst_op1_d=reg_32(static_cast<uint8_t>(inst.rm_index));
 			if (inst.rm<0xc0) SaveMd(inst.rm_eaa,inst_op1_d);	// early write-pf
 			SETFLAGBIT(ZF,1);
 		} else {
